WIP - importing object models
Swap to glb format for the model
Can get diffuse color on material and vertex data from object model

+++ Glaceon/Utils.cpp
@@ -7,15 +7,15 @@ bool CompareGlmVec3(const glm::vec3& a, const glm::vec3& b) {
   return false;
 }
 
-std::vector<size_t> GetIndexFromVertexData(const std::vector<glm::vec3>& vertexData, std::vector<glm::vec3>& uniqueVertex) {
+std::vector<uint32_t> GetIndexFromVertexData(const std::vector<glm::vec3>& vertexData, std::vector<glm::vec3>& uniqueVertex) {
   if (vertexData.empty()) { return {}; }
 
-  std::vector<size_t> indexes;
+  std::vector<uint32_t> indexes;
   for (glm::vec3 vec : vertexData) {
     auto iter = std::ranges::find_if(uniqueVertex, [&](const glm::vec3& v) { return CompareGlmVec3(v, vec); });
     if (iter == uniqueVertex.end()) {
       // New unique vec3 found
-      size_t newIndex = uniqueVertex.size();
+      uint32_t newIndex = uniqueVertex.size();
       uniqueVertex.push_back(vec);
       indexes.push_back(newIndex);
     } else {
