+++ Glaceon/Utils.h
@@ -11,11 +11,11 @@ namespace glaceon {
  * @param vertexData The vector of glm::vec3 vertices.
  * @param uniqueVertex The vector to store the unique vertices.
  *
- * @return A vector of size_t indices corresponding to the unique vertices in the vertexData vector.
+ * @return A vector of uint32_t indices corresponding to the unique vertices in the vertexData vector.
  *
  * @throws None
  */
-std::vector<size_t> GetIndexFromVertexData(const std::vector<glm::vec3>& vertexData, std::vector<glm::vec3>& uniqueVertex);
+std::vector<uint32_t> GetIndexFromVertexData(const std::vector<glm::vec3>& vertexData, std::vector<glm::vec3>& uniqueVertex);
 
 }// namespace glaceon
